[EI-226] [web] refactor: Refactor indicator-related hooks and chat components
