+++ web/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook.ts
@@ -57,6 +57,8 @@ export const useSelectedIndicatorBoardMetadata = () => {
       return;
     }
 
+    console.log('호출!');
+
     addCustomForecastIndicatorTrigger(
       {
         customForecastIndicatorId,
