+++ web/app/ui/components/view/molecule/chat-card/chat-card-content.tsx
@@ -18,9 +18,11 @@ const ChatCardContent = ({ messages = [], isLoading }: ChatCardProps) => {
   return (
     <Collapsible.Content className="h-[35vh] overflow-y-auto	rounded-b-xl border border-fingoo-gray-3 bg-white">
       <div className="space-y-3 p-3">
-        {messages.map((message) => (
-          <MessageItem key={message.id} role={message.role} content={message.content} />
-        ))}
+        {messages.map((message) =>
+          message.role === 'user' || message.role === 'assistant' ? (
+            <MessageItem key={message.id} role={message.role} content={message.content} />
+          ) : undefined,
+        )}
         {isLoading ? <MessageItem role="assistant" content={<DotSpinner />} /> : null}
         <div ref={Chatref}></div>
       </div>
